Added the beginnings of a TargetSpawningMinigame(OSU)
It spawns targets that you have to click on, and it ends the game if 10 targets are on the screen at the same time. The idea is to have it be a status effect that a boss can put on you, so you have to do something ontop of just playing normally or you lose/take big damage.